pyproject.toml
[tool.poetry]
name = "unreasonable-llama-discord"
version = "0.2.0"
description = "Simple Discord bot using unreasonable-llama"
authors = ["SteelPh0enix <wojciech_olech@hotmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
discord-py = "^2.4.0"
unreasonable-llama = "0.2.2"
transformers = "^4.42.3"
jinja2 = "^3.1.4"
sentencepiece = "^0.2.0"
protobuf = "^5.27.2"


[tool.poetry.group.dev.dependencies]
ruff = "^0.5.5"
ruff-lsp = "^0.0.54"
mypy = "^1.11.0"

[tool.ruff]
line-length = 120

[tool.mypy]
disable_error_code = ["type-arg", "valid-type"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
